+++ common/src/main/java/net/mca/Config.java
@@ -28,7 +28,9 @@ public final class Config implements Serializable {
 
     //mod features
     public boolean overwriteOriginalVillagers = true;
+    public List<String> moddedVillagerWhitelist = List.of();
     public boolean overwriteOriginalZombieVillagers = true;
+    public List<String> moddedZombieVillagerWhitelist = List.of();
     public float babyZombieChance = 0.25f;
     public boolean villagerTagsHacks = true;
     public boolean enableInfection = true;