+++ common/src/main/java/net/mca/server/SpawnQueue.java
@@ -10,6 +10,7 @@ import net.minecraft.entity.Entity;
 import net.minecraft.entity.SpawnReason;
 import net.minecraft.entity.mob.ZombieVillagerEntity;
 import net.minecraft.entity.passive.VillagerEntity;
+import net.minecraft.util.registry.Registry;
 
 import java.util.LinkedList;
 import java.util.List;
@@ -74,12 +75,14 @@ public class SpawnQueue {
             return false;
         }
         if (Config.getInstance().overwriteOriginalVillagers
-                && entity.getClass().equals(VillagerEntity.class)
+                && (entity.getClass().equals(VillagerEntity.class) ||
+                    Config.getInstance().moddedVillagerWhitelist.contains(Registry.ENTITY_TYPE.getId(entity.getType()).toString()) && entity instanceof VillagerEntity)
                 && !villagerSpawnQueue.contains(entity)) {
             return villagerSpawnQueue.add((VillagerEntity)entity);
         }
         if (Config.getInstance().overwriteOriginalZombieVillagers
-                && entity.getClass().equals(ZombieVillagerEntity.class)
+                && (entity.getClass().equals(ZombieVillagerEntity.class) ||
+                    Config.getInstance().moddedZombieVillagerWhitelist.contains(Registry.ENTITY_TYPE.getId(entity.getType()).toString()) && entity instanceof ZombieVillagerEntity)
                 && !zombieVillagerSpawnQueue.contains(entity)) {
             return zombieVillagerSpawnQueue.add((ZombieVillagerEntity)entity);
         }